move data types to connector crate
