+++ feagi_connector_core/src/data_types/image_frame.rs
