+++ feagi_connector_core/src/data_types/percentages.rs
