+++ feagi_connector_core/src/data_types/processing/image_frame_processor.rs
