+++ feagi_connector_core/src/data_types/segmented_image_frame.rs
