+++ feagi_connector_core/src/neuron_coding/xyzp/coder_traits.rs
