+++ feagi_connector_core/src/neuron_coding/xyzp/decoders/misc_data_absolute.rs
