+++ feagi_connector_core/src/neuron_coding/xyzp/decoders/percentage_4d_fractional_exponential.rs.disabled
