+++ feagi_connector_core/src/neuron_coding/xyzp/decoders/percentage_fractional_exponential.rs.disabled
