+++ feagi_connector_core/src/neuron_coding/xyzp/encoders/image_frame.rs
