+++ feagi_connector_core/src/neuron_coding/xyzp/encoders/mod.rs
