+++ feagi_connector_core/src/neuron_coding/xyzp/encoders/percentage_linear.rs.disabled
