+++ feagi_connector_core/src/neuron_coding/xyzp/encoders/segmented_image_frame.rs.disabled
