+++ feagi_connector_core/src/neuron_coding/xyzp/encoders/signed_percentage_fractional_exponential.rs.disabled
