+++ feagi_connector_core/src/neuron_coding/xyzp/mod.rs
