+++ feagi_data_structures/src/common_macros/cartesian.rs
