+++ feagi_data_structures/src/common_macros/mod.rs
