--- feagi_data_structures/src/data/mod.rs
@@ -1,12 +1,0 @@
-mod feagi_json;
-mod image_frame;
-mod segmented_image_frame;
-pub mod descriptors;
-mod misc_data;
-mod percentages;
-
-pub use image_frame::ImageFrame;
-pub use segmented_image_frame::SegmentedImageFrame;
-pub use feagi_json::FeagiJSON;
-pub use misc_data::MiscData;
-pub use percentages::*;
