+++ feagi_data_structures/src/lib.rs
@@ -1,14 +1,13 @@
-pub mod basic_components;
-pub mod data;
-pub mod wrapped_io_data;
-pub mod genomic;
-pub mod neurons;
-pub mod processing;
 mod templates;
 mod error;
 mod feagi_signal;
+mod feagi_json;
+pub mod common_macros;
+pub mod genomic;
+pub mod neurons;
 
 pub use templates::*;
 pub use error::FeagiDataError as FeagiDataError;
 pub use feagi_signal::{FeagiSignal, FeagiSignalIndex};
+pub use feagi_json::FeagiJSON;
