+++ feagi_data_structures/src/neurons/xyzp/mod.rs
@@ -1,10 +1,8 @@
 mod neuron_xyzp;
 mod neuron_xyzp_arrays;
 mod cortical_mapped_xyzp_neuron_data;
-mod coders;
+
 
 pub use neuron_xyzp::NeuronXYZP;
 pub use neuron_xyzp_arrays::NeuronXYZPArrays;
 pub use cortical_mapped_xyzp_neuron_data::CorticalMappedXYZPNeuronData;
-
-pub use coders::{NeuronXYZPEncoder, NeuronXYZPDecoder, encoders, decoders};
